rename: filter list to filter net
